requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Auth
python-jose[cryptography]==3.3.0
authlib==1.3.0
httpx==0.26.0
itsdangerous==2.1.2

# FastAPI security
python-multipart==0.0.6

# Config
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Logging
structlog==24.1.0

# Dev
pytest==7.4.4
pytest-asyncio==0.23.3
